// src/lib/api/gen/types/BillingRequestResponseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { WorkflowUserResponseDto } from "./WorkflowUserResponseDto.ts";

export type BillingRequestResponseDtoStatusEnum =
  | "DRAFT"
  | "SUBMITTED"
  | "UNDER_REVIEW"
  | "REJECTED"
  | "APPROVED"
  | "INVOICED"
  | "CANCELLED";

export type BillingRequestResponseDto = {
  /**
   * @type string
   */
  id: string;
  /**
   * @type string
   */
  requesterId: string;
  /**
   * @type object
   */
  requester: WorkflowUserResponseDto | null;
  /**
   * @type string
   */
  departmentId: string | null;
  /**
   * @type string
   */
  customerName: string;
  /**
   * @type string
   */
  customerEmail: string | null;
  /**
   * @type string
   */
  customerAddress: string | null;
  /**
   * @type string
   */
  title: string;
  /**
   * @type string
   */
  description: string | null;
  /**
   * @type string
   */
  amount: string;
  /**
   * @type string
   */
  currency: string;
  /**
   * @type string
   */
  billingCategory: string;
  /**
   * @type string
   */
  status: BillingRequestResponseDtoStatusEnum;
  /**
   * @type string
   */
  workflowInstanceId: string | null;
  /**
   * @type string
   */
  invoiceId: string | null;
  /**
   * @type boolean
   */
  canResubmit: boolean;
  /**
   * @type string
   */
  rejectionReason: string | null;
  /**
   * @type object
   */
  customFieldsJson: object | null;
  /**
   * @type string, date-time
   */
  submittedAt: string | null;
  /**
   * @type string, date-time
   */
  approvedAt: string | null;
  /**
   * @type string, date-time
   */
  rejectedAt: string | null;
  /**
   * @type string
   */
  createdAt: string;
  /**
   * @type string
   */
  updatedAt: string;
};